# app/api/dev/dashboard/[gameId]/upgrade/confirm/route.ts
import { NextResponse, type NextRequest } from 'next/server';
import { backendFetch } from '@/lib/backend-auth';

export async function POST(request: NextRequest, { params }: { params: Promise<{ gameId: string }> }) {
  try {
    const { gameId } = await params;
    const body = await request.json();
    const response = await backendFetch(`/v1/dev/dashboard/${gameId}/upgrade/confirm`, {
      method: 'POST',
      body: JSON.stringify(body),
    });

    if (response === null) {
      return NextResponse.json({ error: 'Upgrade confirmation not found' }, { status: 404 });
    }

    return NextResponse.json(response);
  } catch (error) {
    return NextResponse.json({ error: error instanceof Error ? error.message : 'Failed to confirm upgrade' }, { status: 500 });
  }
}