# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TurnKit API",
    "description": "TurnKit game backend — multi-tenant leaderboard, matchmaking, relay and player store for indie games.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.turnkit.dev",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/dev/dashboard/{gameKeyId}/upgrade/confirm": {
      "post": {
        "tags": [
          "Dev — Dashboard"
        ],
        "summary": "Confirm upgrade",
        "description": "Confirms an upgrade after the frontend shows the proration preview",
        "operationId": "confirmUpgrade",
        "parameters": [
          {
            "name": "gameKeyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradeConfirmResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dev/dashboard/{gameKeyId}/upgrade/checkout": {
      "post": {
        "tags": [
          "Dev — Dashboard"
        ],
        "summary": "Create upgrade checkout",
        "description": "Creates a checkout for new subscriptions or a proration preview for existing subscriptions",
        "operationId": "createUpgradeCheckout",
        "parameters": [
          {
            "name": "gameKeyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradeCheckoutResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "UpgradeCheckoutRequest": {
        "type": "object",
        "properties": {
          "selectedCcu": {
            "type": "integer",
            "format": "int32"
          },
          "selectedModules": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "LEADERBOARDS",
                "PLAYER_STORE",
                "MATCHMAKING",
                "TURN_RELAY"
              ]
            },
            "minItems": 1
          }
        },
        "required": [
          "selectedCcu",
          "selectedModules"
        ]
      },
      "UpgradeConfirmResponse": {
        "type": "object",
        "properties": {
          "flowType": {
            "type": "string"
          },
          "optimistic": {
            "type": "boolean"
          },
          "subscriptionUpdateRequested": {
            "type": "boolean"
          },
          "checkoutUrl": {
            "type": "string"
          },
          "checkoutSessionId": {
            "type": "string"
          },
          "subscriptionId": {
            "type": "string"
          },
          "targetProductId": {
            "type": "string"
          },
          "targetPlanId": {
            "type": "string"
          },
          "targetPriceInCents": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "UpgradeCheckoutResponse": {
        "type": "object",
        "properties": {
          "checkoutUrl": {
            "type": "string"
          },
          "checkoutSessionId": {
            "type": "string"
          },
          "subscriptionUpdated": {
            "type": "boolean"
          },
          "flowType": {
            "type": "string"
          },
          "requiresConfirmation": {
            "type": "boolean"
          },
          "subscriptionId": {
            "type": "string"
          },
          "currentPlanId": {
            "type": "string"
          },
          "currentPlanName": {
            "type": "string"
          },
          "currentPriceInCents": {
            "type": "integer",
            "format": "int32"
          },
          "targetPlanId": {
            "type": "string"
          },
          "targetPlanName": {
            "type": "string"
          },
          "targetPriceInCents": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedProrationCents": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedChargeTodayCents": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedCreditTodayCents": {
            "type": "integer",
            "format": "int32"
          },
          "billingPeriodStart": {
            "type": "string",
            "format": "date-time"
          },
          "billingPeriodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "estimateIsExact": {
            "type": "boolean"
          },
          "targetProductId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "Client Key or Player JWT"
      }
    }
  }
}